engine: findbugs: Handling static members
This patch removes the general exclusion of FindBugs warnings from the
MS category (various warnings regarding improper handling and security
of static members), in order to prevent future patches from introducing
these potentially dangerous and hard-to-find bugs.

Wherever possible, the pre-existing warnings were fixed by altering the
members' modifiers (usually reducing their visibility, which is a good
defensive technique regardless).
Where it was not possible to EASILY solve the issue a specific exclusion
was added in an attempt not to "rock the boat".

Change-Id: I1b940a5e0e25ae4c9488eb32250dc20700d6cd25
Signed-off-by: Allon Mureinik <amureini@redhat.com>