core, engine: Limit the number of SCSI and VirtIO blk devices
This change modifies the checkPciAndIdeLimit method to verify if the
VM being created or updated respects the maximum number of SCSI devices
that can be attached to both VirtIO SCSI and sPAPR VSCSI controllers.

The maximum allowed number of PCI devices is retrieved from the osinfo
in order to allow a larger value in architectures that have more
available virtual PCI slots than x86-64.

Change-Id: Ibd923ca7b48b65f700a1c7cf6d0d938f12ebe48c
Signed-off-by: Vitor de Lima <vitor.lima@eldorado.org.br>
Signed-off-by: Gustavo Pedrosa <gustavo.pedrosa@eldorado.org.br>