tools: notifier: Remove auxiliary email address from EventSender interface
Prior to this change the interface EventSender.java
contained the following logic: an email address can come from wither
The subscriber table(event_subscriber) or if it does not exist the user
table (users). This patch moves this logic from the interface to the
component loading the email subscriber, as this logic is email specific.

Simpler interfaces are better.

Change-Id: Ic69ff7541fca8347fb7cb85665934ca704fa9d06
Signed-off-by: Mooli Tayer <mtayer@redhat.com>