dbscripts: fix password field length
the password column in the 'storage_server_connections' table has
a different type with respect to the corresponding parameter of
the Insertstorage_server_connection function.

This trigger SQL exception doing the following:

- configure an ISCSI data domain (bug triggered on the first one,
  but it should behave the same with the others too).
- in the 'New domain' window, issue fill all the fields,
  most notably username and password, and do a manual discovery.
- try to add the target which shows up as result of the discovery.

This patch fixes this problem by enlarging the password field on
the table, in order to let it hold the encrypted password.

Bug-Url: https://bugzilla.redhat.com/1052839
Change-Id: Ibe5dcb58dd445216b43bbce63b342b96d5e225e6
Signed-off-by: Francesco Romani <fromani@redhat.com>