gluster: fix remove host when host is not part of cluster
VDSM in hosts with compatitbility version 3.3(or less)
doesn't return appropriate error status(GlusterHostIsNotPartOfCluster)
during the host remove if the operation is failed because
the host which is being removed is not part of the gluster-cluster.

The Host can be present in engine but not in the actual
gluster-cluster if the host installation is failed.

This issue is now fixed by verifying the peer list and
remove the host if the host is not found in the list.

Change-Id: I785beb895f0fb72bc486055c666936073b5a5d59
Bug-Url: https://bugzilla.redhat.com/1050135
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>