packaging: build: devenv: remove files installed previosly
with increasing changes in packaging, and people that preserve their
prefix, we need a solution to cleanup old installed files if are missing
from new installation to avoid conflicts.

a method of storing a file list of all simple files (these that are
copied recursively) is a simple and efficient for this cause.

files are relative to PREFIX so that if people move PREFIX to other
location we won't erase wrong files.

Change-Id: I2b0f47ee6fa4b3fb6759bc4f4899092f4128ced7
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>