engine: Provide defaultRoute property for management network
The engine should pass for setup networks requests a new
property for the management network if it was configured
as DHCP or with a static IP and a gateway was set for it.
The new property will be supported from cluster levels 3.4
and above.

This will allow vdsm to be name-agnostic in regards to the
management network.

Change-Id: Id7bc20065ce28926ceb977c09f0480b345c66da5
Bug-Url: https://bugzilla.redhat.com/1015009
Signed-off-by: Moti Asayag <masayag@redhat.com>