engine: Block sync of network with QoS when not supported
When a QoS entity is attached to network, the network is supposed to
appear as out-of-sync on a host, no matter the cluster compatibility
version of the host. However, when the host's cluster doesn't support
Host Network QoS, it shouldn't be possible to then synchronize the
network (which whould also apply QoS to it).

Change-Id: Ie79c6463676b82f1743fb162fc6c660dd88756f8
Bug-Url: https://bugzilla.redhat.com/1054320
Signed-off-by: Lior Vernia <lvernia@redhat.com>