Maintenance mode shall persist after reboot
Fencing operations can occur as a result of manual operation or
automatically after Host was in non-responsive state and has been
restarted after waiting the grace-time for it to reconnect.

In the case that Host was in Maintenance
1) If it was manually started , it should stay in maintenance
2) If it was manually restarted it should change first to reboot and move back to
Maintenance when the Host is ON again.

In the case that Host is restarted from the non-responsive treatment, it
should go to reboot -> non responding -> up as before

This patch address 1) and 2) above by

1) Setting the parent command in VdsNotRespondingTreatmentComman
parameters so the Restart command can distinguish between manual Restart and
automatic restart
2) Add a flag to FenceVdsActionParameters to mark if Host should be
moved again to Maintenance after it was restarted
3) Use The Host last status to skip changing the status in case of 1)
4) Use the flag described in 2) to set Host again to Maintenance in
case that Host was manually rebooted while in the Maintenance status.

Change-Id: I3dcb06b18c4af13876d1fb3043fb5b6cc5c6a74a
https://bugzilla.redhat.com/show_bug.cgi?id=917039
Signed-off-by: Eli Mesika <emesika@redhat.com>