1. core, tools: Supporting change password Url presentation
This patch adds the ability for the admin to configure for each
authentication domain a url which will be shown on login attempts
failing due to password expiration.

This allows providing the users with a link to a web page allowing to
change their expired password.

In order to set these URLs, a new optional parameter was introduced to
manage-domains, named changePasswordUrl.

Change-Id: I8eb0f858e26bdefffe526623d025fac47791711e
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>