core: Enable ServletUtilsTest.testCanReadFile()
This test was disabled in commit
efa1582eafeac05b5f911dd2ce93fdfe8aa0e542 as it fails when run as root.

This patch offers a cleaner way to make sure the build passes - instead
of ignoring the test, Assume is used to make sure the user is not root,
and if it is, silently skip the test.

Using this technique we can make sure that the test is run at least some
of time, and gain some extra stability.

Change-Id: I754a358221d9fc370d6418a06baf08b97c25b4fa
Signed-off-by: Allon Mureinik <amureini@redhat.com>