core: Enable AuditLogDAOTest.testLongMessageSave()
AuditLogDAO.testLongMessageSave() has been @Ignored since it consistently
fails.
This patch modifies it to stop relying on the new record's generated ID,
so it can conceivably pass, in a similar fashion to testSave().
Additionally, the newly inserted AuditLog was made external (since
internal messages do not get truncated) and the MaxAuditLogMessageLength
config value was introduced to the fixtures database since the
underlying InsertExternalAuditLog stored procedure relies on it, and
VdcOptionDAOTest was amended accordingly.

Change-Id: I30189046c8ad97256717e3c804996a22bf1b5d9b
Signed-off-by: Allon Mureinik <amureini@redhat.com>