restapi: Preserve labels when setup networks is sent
The setup networks api doesn't use the patch
mechanism to preserve any interface's properties that
wasn't sent by the user. Instead, it creates a new
interface entity and set only the provided nic attributes.
Since labels are maintained as a sub-collection on the
nic entity, it should not be deleted for each setup networks
request.

This logic cannot reside on the engine side, since there will
be no distinction between setup networks sent by the UI, in
which the labels can be provided or removed, to the rest in
which labels cannot be provided and should be preserved.

Change-Id: Ic2b728a4797d03d6718e6c9b0b6adec6dcf3bbff
Bug-Url: https://bugzilla.redhat.com/1055441
Signed-off-by: Moti Asayag <masayag@redhat.com>