core: Add storage types to DiskImage entity
Currently DiskImage has a storageIds field that contains a list of the
storages it resides on. It makes sense also to add a storage types field
to avoid having to fetch the storage domain in order to know which storage
type the disk resides on.
This will be helpful when the storage pool storage type property will be
removed.

Change-Id: I8302972bb21fa076cd35fb308f5f660ca9200daa
Relates-To: https://bugzilla.redhat.com/1038053
Signed-off-by: Tal Nisan <tnisan@redhat.com>