core: HttpConnManager - add explicit connection timeout
This patch adds http connection timeout to the connections manager used
to initiate http connections - the connection timeout specifies a
timeout to be used until connection is established. Currently we pass
the timeout to the XmlRpcClientConfig only, therefore we can't count on
it that when we'll use the HTTP connection outside of the XML-RPC
transport layer timeout won't be encounterd.

Change-Id: I4e927ce28c8fdd9bcdcdb37e23b8d45c7515fd55
Signed-off-by: Liron Aravot <laravot@redhat.com>