core: refactor failure to run vm flow - part 2
When migration fails and the VM get back to UP state in the source host,
we do not want to treat the VM as VM that failed to run. Thus,
MigrateVmCommand overrides RunVmCommandBase#failedToRunVm method and do
not do anything in that case.

Instead of override the whole RunVmCommandBase#failedToRunVm method, it
is better to override just the RunVmCommandBase#processVmPoolOnStopVm
method so the rest of the things that failedToRunVm do would be
executed.

In addition, the code in RunVmCommandBase#reportCompleted was
simplified a bit.

Change-Id: I6dc8e13f1ab932283557b3cbf16a34e72b9b3a69
Signed-off-by: Arik Hadas <ahadas@redhat.com>