core: fix possible NPE on migration failure
We recently added a call to decrease pending memory from the destination
VDS when migration ends. This code was added to the
MigrateVmCommand#reportCompleted method and we assume that the
destination VDS is not null when we reach that method, but in some cases
where the migration fails we might reach the reportCompleted method
after the destination VDS is set to null (see MigrateVmCommand#rerun).

This patch solves this issue by adding null-check which ensures that the
destination VDS is valid before calling to decrease its pending memory.

Change-Id: If5c975ba5a0824c80be36cb997b3b040b5191192
Bug-Url: https://bugzilla.redhat.com/1048790
Signed-off-by: Arik Hadas <ahadas@redhat.com>