core: Execute clear tasks with new thread.
clearAsyncTasksWithOutVdsmId is executed at the finally step when
executing a vdc command.
If the execute phase encountered a problem, the engine sets the
transaction to rollback only and we can not use it any more.
Since the clearAsyncTasksWithOutVdsmId uses calls to the DB, the engine
throws an EJBTransactionRolledbackException, and the original exception does not
seep to the user.

The proposed solution is to clear the async tasks with a new
transaction, so it will not be dependent on the related execute phase.

Change-Id: I3434c2dbd605e741b2cfbbede8c4a120f21ca9fe
Bug-Url: https://bugzilla.redhat.com/1037441
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>