gluster: check gluster daemon status in sync-job
In sync-job, peer list will be retrieved always from the first
UP server. There are chances that gluster daemon in other servers
may be down. In this case, peer status for those servers will DISCONNECTED.

When a host is found in UP state in Database and peer status is
'DISCONNECTED', that host will be moved to Non-Operational status.
'peer status' command will be executed on that server to make sure
the gluster daemon is actually down before moving to Non-Operational.

Change-Id: I3e0f661cb496e741f7c06df42ce3b55037a87e28
Bug-Url: https://bugzilla.redhat.com/1056997
Bug-Url: https://bugzilla.redhat.com/1097736
Signed-off-by: Kanagaraj M <kmayilsa@redhat.com>
Signed-off-by: Sahina Bose <sabose@redhat.com>