engine: Fix NPEx vulnerabities related to osinfo change
This patch fixes couple of NPEx vulnerabities related to ConsolesBase
and OS info in AsyncDataProvider.

The vulnerabities happened when changes in osinfo caused invalidity of
VMs already stored in the database.

The problem was fixed by introducing null checks in appropriate places
and a map with the possibility of returning default value if requested
key is not found.

Change-Id: I04713a90374cac30fd358fdd0fcf02d2a020d477
Signed-off-by: Frantisek Kobzik <fkobzik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1046809