restapi : ints and shorts overflow
int and short values are overflown before
they even reach the business code, allowing
for wrong values to be set.

When posting to rest api, if the value of an
int is greater than Integer.MAX_VALUE or value
of a short is larger than Short.MAX_VALUE. The
values are overflown when values are unmarshalled
using JAXB.

An excpetion should be raised during JAXB
unmarshalling to indicate that the value is
greater than the MAX_VALUE for the type.

Change-Id: I9d0be6f3f220d7a5f58bb47906a8565d80e0ebf0
Bug-Url: https://bugzilla.redhat.com/1000796
Signed-off-by: Ravi Nori <rnori@redhat.com>