core: Don't duplicate directory name in the DB
Currently in the "users" table we store the directory name in the
"domain" column and also, as a suffix added to the user name, in the in
the "username" column. This duplication causes confusion, and makes the
SQL schema a bit more non normalized. This patch removes that
duplication and changes the places that need to display the directory
name (GUI and RESTAPI) so that it is done explicitly.

Bug-Url: https://bugzilla.redhat.com/1055881
Bug-Url: https://bugzilla.redhat.com/1059258
Change-Id: I53ebaf3254f829d73b2574230ae333f13cfcd546
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 2d36ccaecbf6a83340525083980f98ddff064e48)