engine: refactor: add Model attribute for help tagging
1. Simple refactor to add a Model attribute for storing a help tag. Previously,
the dialog hashname was used to achieve help URL mapping to dialogs. Help mapping
deserves its own attribute, so we now use the helpTag attribute to achieve the mapping.

2. Added HelpTag enum to group all help tags into one file. Each help tag has a name
and comment. The name is used in the help-mapping-file lookup. The comment
is not used by GWT code -- it is simply a helper for documentation writers to know where
to find a help tag in the application (i.e. where is the dialog). helptag.py will add the
comment to the help-mapping-file it generates -- see below.

3. Created a utility script, helptag.py, to scan the HelpTag file for help tags
and generate the help-mapping-file. The main purpose of this tool is to help find changes
in the help tags around release time so new help URLs can be added to the help mapping
files as needed.

helptag.py can generate a simple diff, or it can generate an entire template (that could then be
diffed with any existing mapping file).

Change-Id: Ia4074fcc2ecfcbdd2ea6c0855d92f2aa4bd26a5b
Bug-Url: https://bugzilla.redhat.com/1014859
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>