5. core: Supporting custom message to appear in case of expired password
The bug description stays that in case of failed login,
the user may get a URL to change the passord, or a custom message.
In order to support that, the following changes were introduced:
a. manage-domains code  -changePasswordUrl was changed to -changePasswordMsg.
b. The input is done in interactive mode.
c. The read value is encoded (as may contain characters like space, : and ,)
d. The manage-domains script was changed accordingly
e. The vdc_options relevant entry was changed (so was the ConfigValues relevant enum literal).
f. ProsivinalAuthenticationResult reads the entry from the configuration and decodes it.
g. A proper VDC BLL message is used depending on whether the content is URL or not (a new
VdcBllMessage was introduced).

Bug-Url: https://bugzilla.redhat.com/1053601

Change-Id: I9f656f38f7f48e169ed01ad084d7f424e5c749f3
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>