tools: Provide better explaination on why setting password failed.
Instead of providing the notes, a better error message is printed:
In case the admin-pass-file does not exist, a message indicating the
file does not exist will be printed.
In other cases, if the value is not "interactive" and the file does
not exist, a short message explaining the problem will be presented.

Bug-Url: https://bugzilla.redhat.com/1058957
Change-Id: I3cbfb4d8ff6cd3dc3a6bdb0b6664d5cfc8012b1a
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>