restapi: Validate unsigned short integers correctly
Recently, in commit 14e9da, we introduced a validation of integer values
provided to the RESTAPI in order to avoid overflows of Java integer
types. This validation doesn't work correctly for unsigned short
integers, as it compares them to the maximum signed short integer. This
has the side effect of rejecting requests that include port numbers
greater than 32767. A notable example is the VDSM port number 54321.
This patch changes the validation so that it will validate unsigned
short integers correctly, comparing them to 65535.

Change-Id: I49e64d5f55c32463950e648aa1cd8c3084017da6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>