core: backward compatibility for monitor devices
Before v3.1 we used to have just one item for monitors in the OVF file
which contained an element called VirtualQuantity that represented the
number of monitors the VM has.

Now we're setting monitor device per monitor item in the OVF file,
assuming that we have a separate item for each monitor. Since it doesn't
hold for OVF that was created in versions which are earlier that 3.1,
when we import a VM that was exported in version 3.0 and below, it will
always have just one monitor device.

The solution is that in case we import a VM that was exported in version
which is early than 3.1, the number of monitor devices that will be
created will be according to the VirtualQuantity value.

Change-Id: I00a88f7856bc7f356f3ac5fd6eac8e432f659d09
Bug-Url: https://bugzilla.redhat.com/1060705
Signed-off-by: Arik Hadas <ahadas@redhat.com>