restapi: Empty String Validation (#1005296)
The generic input validation method in the API - validateParameters() uses reflection
to invoke isSetX() for the given field. In the particular case of when the filed is
a string, .isEmpty() check should be added. This is what this patch does.

This fix is done in the context of a specific bug - 1005296 - which complains
about empty <address> field passing validation when creating a new storage
server connection

Bug-Url: http://bugzilla.redhat.com/1005296
Change-Id: Ib0556725ad2ddc607d587c079bad451e4b74972d
Signed-off-by: Ori Liel <oliel@redhat.com>
(cherry picked from commit 0107892b7c0b3a6c0ccad0da0c077a1966fc1f8a)