aaa: creating extension-manager module
This module will hold all required classes for loading and
managing extensions. Currnently this contains the
Manage, Factory, and the Configuration related classes.

The patch deals with the following:
a. Moving these classes to the new module and fixing imports all over the code accordingly
b. Fixing the MANIFEST.MF of the ear to add dependency on the new module.
c. Fixing the get-extension to compile against authentication module as well,
as currently DirectoryManager code resides there and needed for custom serialization
of DirectoryUser and DirectoryGroup
d. Adding a module dependency for authentication - to be dependent on extension-manager

Topic: AAA
Change-Id: I3c7258206880fd974934672de9fb5900192cca33
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>