core: Reorganize SD space validations in AddVm
This is the first of two patches meant to fix faulty storage space
allocations validations in the different add VM commands.
This patch keeps the current (faulty) validations and reorganizes them
in methods that can be overridden in the derived classes when needed.
It also adds two new threshold tests - verifying the AddVmCommand
behaviour to the simulated
StorageDomainValidator.isDomainWithinThresholds .

Change-Id: I482709d7edb0333a6aa7b9ab723eea7f4c5a00d3
Related-to: https://bugzilla.redhat.com/show_bug.cgi?id=917682
Related-to: https://bugzilla.redhat.com/show_bug.cgi?id=1053742
Signed-off-by: Vered Volansky <vvolansk@redhat.com>