core: Make VM disks editable
When a VM is updated the data from the database is compared to the
updated VM object provided by the caller. But the VM object loaded from
the database doesn't contain the list of disks, so if the caller
provides an VM object that contains the list of disks the UpdateVM
command will wrongly assume that the caller is trying to modify this
list. To avoid this issue this patch marks the list of disks as
editable, so that the UpdateVM command will ignore it.

Change-Id: I0fa0a0397a44adc93680eae593a92346f9e67600
Bug-Url: https://bugzilla.redhat.com/1059713
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>