core: Fix storage validations for extending a disk
Use StorageDomainValidator.hasSpaceForNewDisk(DiskImage) method, which
is sensitive to the type and format of the disk. Although a new disk is
NOT added, the extension is essentially empty, so sans some overhead, it
can be considered as a new disk.

This patch includes fixing the command and adding some basic unit tests
to the flow.

Change-Id: I76d38c41321495b9e12087dc4698f67830d7a1d4
Bug-Url: https://bugzilla.redhat.com/1053740
Signed-off-by: Allon Mureinik <amureini@redhat.com>