core: fix vm pools behavior when DC is missing
When DC is missing (for example force-removed)
then if any vm pools still exist, they don't return in queries, because
of bad join in the view.
 this was fixed by changing the vm_pools_view to use left-join with
storage-pools

in addition, this should not happen, as empty vm pools are not
supported.
but when using force-remove DC, all vms are deleted, but vm-pools are
left.
 this was fixed by changing the stored procedure for remove storage
pool, to remove vm-pools as well, after removing vms.

Change-Id: I47c963b6b66efa05e96cb9b9f80e034d8070a30b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1031717
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>