core: refactor failure to run vm flow - part 4
Remove MigrateVmCommand#failedToMigrate which was confusing now that we
have MigrateVmCommand#runningFailed method.

In addition MigrateVmCommand#runningSucceeded is removed.

For the changes above, the calls to CommandBase#freeLock was moved to
the relevant places in RunVmCommandBase, which are better places for
them as they are relevant to all command that extend RunVmCommandBase.

Note that the freeLock method handle the case where the lock is null,
so it is ok in case the command release the lock at the end of the
execute method as well.

Change-Id: I42cefc4c5f7974730a2b38f63f310ab975f73622
Bug-Url: https://bugzilla.redhat.com/1049321
Signed-off-by: Arik Hadas <ahadas@redhat.com>