core: AddDiskCommand - add option to leave disk as illegal on failure
When adding a disk in case of failure the added disk is currently being
removed automatically by the engine which is wrong in atleast some of
the cases as we might have garbage on the storage.
As first step, this change allows to restrict the command to leave the
disk in illegal status in case of failure and not delete the disk.

Change-Id: I8ccf85a2ab38e96dc934be36acb5ac70a41f5e47
Signed-off-by: Liron Aravot <laravot@redhat.com>