userportal, webadmin: fix 404 regression for help files
Patch http://gerrit.ovirt.org/#/c/21052/ introduced a regression where if no
help mapping json files are found, a 404 is served to the UI and logged.

This fixes that by sending back an empty JSON string.

There was also a path mapping bug in patch http://gerrit.ovirt.org/#/c/21052/
and this fixes that as well.

Change-Id: I219e986ac45e7a05801bd9b519a006a429e5f82f
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>