webadmin: CustomPropertyValidation blank string
Fixed up the way a blank string is tested.
Before this patch, it's trim()ed and compared to "" using
StringHelper.stringEquals. Since trim() is called on this string, we can
assume it's not null, so StringHelper.stringEquals is redundant. Since
we know we have a non-null instance, we can simply used the optimized
isEmpty().

An additional advantage is removing another call to
StringHelper.stringsEqual, moving one step closer to eliminating this
method.

Change-Id: I3592a500d5af41593a3f8f9eb1e888d6ac22cf04
Signed-off-by: Allon Mureinik <amureini@redhat.com>