webadmin: VmGeneralModel empty string validation
Replaced the test for domain's emptiness from
StringHelper.stringsEquals(vm.getCustomProperties(), "") to a simpler
StringHelper.isNullOrEmpty(vm.getCustomProperties()). Note that
although there is a slight difference in semantics, there is no
point in having null custom properties, and it's implicitly assumed
to be not-null in other places of the code.

Change-Id: Ib45cf6fb428cd6750a1264ca45a8b332420bc57f
Signed-off-by: Allon Mureinik <amureini@redhat.com>