core: Prevent StopVm to interleave with other VM locked actions
Stop and Shutdown VM commands are calling the DestoryVds command which is contending on
the VdsManager lock. The lock is implemented as synchronized (object) -
this means that order isn't kept and we're subjected to JVM's scheduling
descisions so other commands, like RunVm can take the lock and result the Vm to stop instead of start

Fix: take an exclusive engine lock on Stop and Shutdown commands

This will also prevent a situation where asyncRunningVms has an entry
although the VM is down.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1038351
Change-Id: I6109ef0ea4869b59a93be6fef4b0c3cff50da62e
Signed-off-by: Roy Golan <rgolan@redhat.com>