core: Allow manual fence in connecting state
Allow to manually fence a Host that is in Connecting state.
Up to now, when host had a network problem, it was switched to
Connecting state for 3 attempts, after those attempts, if host was still
unreachable, the host was turned to non-responding status.
During the Connecting status, if you would try to manually fence the
host, you would get an error saying:
"Due to intermittent connectivity to this Host, fence operations are not
allowed at this time. The system is trying to reconnect, please try
again in 30 seconds."

If an outage problem occurs, this limitation forces us to wait for all
3 network reconnect attempts before we can confirm that the host was
rebooted and perform a manual fence (which can take several minutes)

This patch removes this limitation and allow manually fencing a host
that is in the Connecting state.

Change-Id: I37a63d299244bc88a35e9c264e696ac88e6d7ef0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1066400
Signed-off-by: Eli Mesika <emesika@redhat.com>