engine: Initialize MAC pool partially
If MAC pool exceeds the maximum then we will initialize it until its
full. An error is being written to the log in this case.

The partial initialization will prevent an OOM in case of a really big
range (not entirely though, this will be fixed in the future).

Change-Id: I0547509eb81a51599e2342503c9f108df1fa47c7
Bug-Url: https://bugzilla.redhat.com/1065011
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
(cherry picked from commit ee37b7289f9d1a6c2b803a3650995c4127876131)