core: decrease pending memory & cpu from the right vds on run vm
On rerun attempt in RunVm/RunVmOnce flow we could decrease the pending
resources from the wrong VDS, i.e not from the VDS the VM was supposed
to run on, in case dedicated VDS was defined for the VM or a specific
VDS was set in RunVmOnce flow.

This patch changes the selection of the VDS for which we should decrease
the pending resources to use RunVmCommandBase#getCurrentVdsId method
which should always return the VDS on which the VM was about to run on.

Change-Id: I7b82e4ef4f9e5ed60542e756640861f7403be090
Bug-Url: https://bugzilla.redhat.com/1061811
Signed-off-by: Arik Hadas <ahadas@redhat.com>