core: remove RunVmCommandBase#getDestinationVds
In light of 84fc3c7 RunVmCommandBase#getDestinationVds is now used
only by classes which extend RunVmCommandBase and is used differently:
MigrateVm returns the VDS that the VM is going to migrate to, RunVm &
RunVmOnce return the predefined VDS on which the VM should run on.

This patch removes RunVmCommandBase#getDestinationVds method. In RunVm
it is replaced by getPredefinedDestinationVdsId method which reflect the
result better by its name and returns only the ID of the VDS.

In addition, minor improvement was made in
MigrateVmCommand#getDestinationVds, and calls to it where we only need
the ID of the VDS were replaced by calls to
MigrateVmCommand#getDestinationVdsId.

Change-Id: Ie1d8fb852a88203378b3bcd3f1d6f96e469e5cb9
Signed-off-by: Arik Hadas <ahadas@redhat.com>