restapi: Duplicate name 'nics' in schema
QA tools assumes unique element names in the schema (api.xsd),
so naming inner element 'nics', which outer element by this name
already exists in the schema causes failure. Again, the failure
is in the QA tool, and is not a real code or business faliure.
Nevertheless, it's enough of a reason to change the element's name.

Bug-Url: http://bugzilla.redhat.com/1063187
Change-Id: Ibe88dc0fd9129776bf462921b42849ba99e12644
Signed-off-by: Ori Liel <oliel@redhat.com>