core: Support detach Storage Domain containing entities.
Adding support for detaching a Data Storage Domain which contains disks, VMs
or Templates.
We should not allow detach of Storage Domain if it contains VM from
pools or VM in preview.
When detaching a Storage Domain, disk snapshots/shareable disks will not be saved
since they are not contained in the OVF.

This patch is part of the feature ImportStorageDomain
(see http://www.ovirt.org/Features/ImportStorageDomain)

Change-Id: I971fe6acd4a2667a09487c5e1108cf7c759587f1
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>