engine: added HelpTagType to HelpTag
Added a HelpTagType enum, and references to it in HelpTag. It's not used in GWT code,
but helptag.py can now use it to filter down its ini/diff files to only those
tags required specifically by webadmin or userportal.

Updated helptag.py to use the HelpTagType to filter the help tags it scans.

Change-Id: I64a3702878b9d341d78119585b51129478bb6d37
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>